[vim] Override statusline from the default qf ftplugin
Vim's quickfix filetype plugin sets the statusline option regardless of
whether the statusline is already set or not; this change overrides the
statusline again for quickfix windows.

An option to prevent the quickfix filetype plugin from setting the
statusline was included in Vim (see
https://github.com/vim/vim/pull/4662), although using this option in
Neovim (untested in Vim) sets the quickfix window's statusline to Vim's
default, which is still undesired behavior.